+++ resources/examples/ts/git.ts
@@ -1,11 +1,14 @@
 
+// @ts-nocheck
+import * as vscode from "vscode";
+
 // Download https://github.com/microsoft/vscode/blob/main/extensions/git/src/api/git.d.ts
 // to the same directory as this macro to enable type definitions. You can use the
 // `Download d.ts files` code action on the link above.
+import { API } from './git';
 
 async function main() {
-  /** @type {import('./git').API | undefined} */
-  const git = vscode.extensions.getExtension('vscode.git')?.exports.getAPI(1);
+  const git = vscode.extensions.getExtension('vscode.git')?.exports.getAPI(1) as API | undefined;
   if (!git) {
     vscode.window.showErrorMessage('Git API not found');
     return;
@@ -17,7 +20,7 @@ async function main() {
     return;
   }
 
-  vscode.window.showInformationMessage(
+  await vscode.window.showInformationMessage(
     `Repository: ${require('path').basename(repo.rootUri.toString())}` +
     `\nBranch: ${repo.state.HEAD.name}` +
     `\nStaged changes: ${repo.state.indexChanges.length}` +